server/requirements.txt
bcrypt
fastapi[standard]
python-dotenv
bcrypt
sqlalchemy